+++ packages/pdfx/lib/src/viewer/interactive_viewer.dart
@@ -91,18 +91,11 @@ class InteractiveViewer extends StatefulWidget {
     this.scrollControls = InteractiveViewerScrollControls.scrollScales,
     this.transformationController,
     required Widget this.child,
-  })  : assert(alignPanAxis != null),
-        assert(child != null),
-        assert(constrained != null),
-        assert(minScale != null),
-        assert(minScale > 0),
+  })  : assert(minScale > 0),
         assert(minScale.isFinite),
-        assert(maxScale != null),
         assert(maxScale > 0),
         assert(!maxScale.isNaN),
         assert(maxScale >= minScale),
-        assert(panEnabled != null),
-        assert(scaleEnabled != null),
         // boundaryMargin must be either fully infinite or fully finite, but not
         // a mix of both.
         assert(
@@ -140,17 +133,11 @@ class InteractiveViewer extends StatefulWidget {
     this.scrollControls = InteractiveViewerScrollControls.scrollScales,
     this.transformationController,
     required InteractiveViewerWidgetBuilder this.builder,
-  })  : assert(alignPanAxis != null),
-        assert(builder != null),
-        assert(minScale != null),
-        assert(minScale > 0),
+  })  : assert(minScale > 0),
         assert(minScale.isFinite),
-        assert(maxScale != null),
         assert(maxScale > 0),
         assert(!maxScale.isNaN),
         assert(maxScale >= minScale),
-        assert(panEnabled != null),
-        assert(scaleEnabled != null),
         // boundaryMargin must be either fully infinite or fully finite, but not
         // a mix of both.
         assert(